Move and namespace `MailingListController` under `GetIntoTeaching`
